[bd-3gh7] Add integration tests for projections crate
Add 24 integration tests across two files:
- cursor_and_rebuild.rs: cursor save/load/resume, idempotent
  try_apply_event, shadow table create/swap, shadow cursor tracking
- digest_and_admin.rs: deterministic digest computation, versioned
  digest staleness detection, admin status/consistency/list queries

Fix NUMERIC→BIGINT cast in admin query_projection_list SQL to avoid
sqlx type mismatch when SUM(events_processed) returns NUMERIC.

+++ platform/projections/src/admin.rs
@@ -239,8 +239,8 @@ pub async fn query_projection_list(pool: &PgPool) -> Result<ProjectionListRespon
     }
 
     let rows = sqlx::query_as::<_, (String, i64, i64, Option<DateTime<Utc>>)>(
-        "SELECT projection_name, COUNT(*) as tenant_count, \
-                COALESCE(SUM(events_processed), 0) as total_events, \
+        "SELECT projection_name, COUNT(*)::BIGINT as tenant_count, \
+                COALESCE(SUM(events_processed), 0)::BIGINT as total_events, \
                 MAX(updated_at) as last_updated \
          FROM projection_cursors \
          GROUP BY projection_name \
